Tests/LibWeb/Text/expected/HTML/message-port-still-delivers-message-even-when-its-immediately-closed.txt

Port 1 received message: Hello world!
